Important Change!!! A new track file format!!!
This new format allows more efficient reading from and writing to multiple
tracks. This change is critical for efficient implementations of several
scripts that are soon to be added, e.g. the rdf script.

+++ lib/tracks/parse.py
@@ -37,7 +37,7 @@ def parse_slice(s):
         if word == '':
             result.append(default)
         else:
-            result.append(float(word))
+            result.append(int(word))
     return slice(*result)
